requirements.txt
discord.py
instaloader
python-dotenv